Ahora Cancelar reservas acepta un parametro 'Motivo' como ya se habia prometido. Se ha desarrollado mas de la mitad del modulo 'Reservas'  en la capa de ASP.Net Core.

+++ SGHR.Application/DTOs/Reservas/ActualizarReservaRequest.cs
@@ -22,7 +22,7 @@ namespace SGHR.Application.DTOs.Reservas
         
         [Required(ErrorMessage = "El numero de huespedes es requerido")]
         [Range(1, int.MaxValue, ErrorMessage = "El numero de huespedes debe ser al menos 1 y no exceder 255")]
-        public sbyte NumeroHuespedes { get; set; }
+        public int NumeroHuespedes { get; set; }
 
     }
 }
